refactor(ws_router): split into 4-file package + LLM-Note headers + sync docs
ws_router.py was a 350-line file mixing WS protocol orchestration with
agent thread + io plumbing. Splitting along the natural boundary makes
each concern locatable and testable in isolation.

Package layout:
  ws_router/
    __init__.py    re-export run_session
    session.py     run_session: read loop, dispatch, cleanup (lifecycle)
    connect.py     handle_connect: auth + merge + reattach trigger
    agent_io.py    start_agent / resume_forwarding / forward_agent_msgs_to_client
                   / _agent_thread_body
    ping.py        ping_loop: 30s keepalive

Renames (drop _ prefix where the function crosses module boundary):
- _handle_connect → handle_connect (connect.py)
- _start_agent → start_agent (agent_io.py)
- _resume_connection → resume_forwarding (agent_io.py — name now reflects
  what it does: it doesn't resume a connection, it spawns a new forward task)
- _forward_agent_msgs_from_io_to_client → forward_agent_msgs_to_client
- _run_agent_thread → _agent_thread_body (still private, used only by start_agent)
- _ping_loop → ping_loop
- io.finish() → io.mark_agent_done()

Inlined (single-caller helpers — premature abstraction):
- _try_runtime_input → INPUT branch in session.py
- _handle_session_status → SESSION_STATUS branch in session.py
- _init_session / _cleanup_session → run_session local vars + finally block

Each new file has a Purpose: / LLM-Note: structured header per project
convention covering Dependencies / Data flow / State / Integration /
Performance / Errors.

Docs synced (7 files):
- protocol.md rewritten around 4-file layout, new function names, inline dispatch
- io.md: io.finish → io.mark_agent_done; _handle_connect → ws_router.connect:handle_connect
- session-reconnect.md / session-websocket-lifecycle.md / client-reconnection.md:
  function name updates throughout
- websocket-protocol.md key-files table updated
- agent-relay-protocol.md: ws_router.py → ws_router/

Public API unchanged: server.py + asgi/websocket.py still
'from .ws_router import run_session'. Tests pass (138).
