+++ docs/network/io.md
@@ -491,11 +491,11 @@ Agent Thread (sync)              Async forwarder / router
 
 `_msgs_from_agent` is append-only. `read_msgs_from_agent` is an async generator that yields events from `self._cursor` onward, advancing the cursor under `_agent_condition` after each batch.
 
-On reconnect, `_handle_connect` calls `io.rewind_to(last_msg_id)` (also under the same lock) to reset the cursor — the new forward task replays everything after that id. If `last_msg_id` is omitted or unknown, cursor rewinds to 0 (full replay; client should dedup by id).
+On reconnect, `ws_router.connect:handle_connect` calls `io.rewind_to(last_msg_id)` (also under the same lock) to reset the cursor — the new forward task replays everything after that id. If `last_msg_id` is omitted or unknown, cursor rewinds to 0 (full replay; client should dedup by id).
 
 ### finish() and close()
 
-- **`io.finish()`** — agent done emitting messages. Sets `_finished` flag and notifies all waiters. `read_msgs_from_agent` returns once it drains remaining buffered events.
+- **`io.mark_agent_done()`** — agent done emitting messages. Sets `_finished` flag and notifies all waiters. `read_msgs_from_agent` returns once it drains remaining buffered events.
 - **`io.close()`** — sets `_closed = True`; subsequent `io.send()` calls become no-ops. Used when the io should accept no more agent output (rare, mostly for shutdown).
 
 There is no sentinel injected into `_msgs_from_client`. A blocked `io.receive()` is unblocked only by an actual client message arriving via `send_to_agent`, or by the encompassing thread being killed at shutdown.
